server/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.3",
    "fastapi>=0.115.0",
    "google-cloud-speech>=2.30.0",
    "httpx>=0.27.2",
    "pymysql>=1.1.1",
    "python-socketio>=5.12.1",
    "sqlalchemy>=2.0.35",
    "uvicorn>=0.31.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "ruff>=0.6.9",
]
